requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Environment variables
python-dotenv==1.0.1

# Computer Vision
opencv-python==4.9.0.80
numpy==1.26.4

# Face Recognition / Deep Learning
deepface==0.0.93
tensorflow==2.15.0
keras==2.15.0
mtcnn==0.1.1
retina-face==0.0.14

# Database
pymongo==4.6.1

# Cloudinary
cloudinary==1.39.0

# HTTP Requests
requests==2.31.0

# Timezone handling
pytz==2024.1

# Utilities
pillow==10.2.0
tqdm==4.66.2
scipy==1.11.4
pandas==2.1.4
